Users

+++ app.py
@@ -1,10 +1,14 @@
-from flask import Flask, render_template, redirect, request
+from flask import flash, Flask, render_template, redirect, request, url_for
 from flask_sqlalchemy import SQLAlchemy
+from flask_login import current_user, LoginManager, login_required, login_user, logout_user, UserMixin
 from flask_migrate import Migrate
 from flask_bootstrap import Bootstrap
 from flask_wtf import FlaskForm
-from wtforms import BooleanField, StringField, SubmitField
-from wtforms.validators import DataRequired
+from werkzeug.security import generate_password_hash, check_password_hash
+from werkzeug.urls import url_parse
+from wtforms import BooleanField, PasswordField, StringField, SubmitField
+from wtforms.validators import DataRequired, Email, EqualTo, ValidationError
+
 import os
 from datetime import datetime
 import sys #just for console printing, can remove after
@@ -20,21 +24,88 @@ app.config['DEBUG'] = True
 db = SQLAlchemy(app)
 migrate = Migrate(app, db)
 bootstrap = Bootstrap(app)
+login = LoginManager(app)
+login.login_view = 'login'
+
+@login.user_loader
+def load_user(user_id):
+    return User.query.get(user_id)
 
 @app.route("/")
 def index():
-	return render_template('index.html')
+	return redirect(url_for('tasks'))
 
 @app.route("/tasks", methods = ('GET', 'POST'))
+@login_required
 def tasks():
-	tasks = Task.query.order_by(Task.last_updated.desc()).all()
+	tasks = current_user.tasks.order_by(Task.last_updated.desc()).all()
 	form = MyForm()
 	if form.validate_on_submit():
-		task = Task(body=form.body.data, deadline=form.deadline.data)
+		task = Task(body=form.body.data, deadline=form.deadline.data, author=current_user)
 		db.session.add(task)
 		db.session.commit()
 		return redirect(request.referrer or url_for('tasks'))
 	return render_template('tasks.html', tasks=tasks, form=form)
+
+@app.route("/login", methods=['GET', 'POST'])
+def login():
+    if current_user.is_authenticated:
+        return redirect(request.referrer or url_for('index'))
+    form = LoginForm()
+    if form.validate_on_submit():
+        user = User.query.filter_by(username=form.username.data).first()
+        if user is None or not user.check_password(form.password.data):
+            flash('Invalid username or password.')
+            return redirect(url_for('login'))
+        login_user(user, remember=form.remember_me.data)
+        next_page = request.args.get('next')
+        if not next_page or url_parse(next_page).netloc != '':
+            next_page = url_for('index')
+        return redirect(next_page)
+    return render_template('login.html', form=form)
+
+@app.route('/register', methods=['GET', 'POST'])
+def register():
+    if current_user.is_authenticated:
+        return redirect(url_for('index'))
+    form = RegistrationForm()
+    if form.validate_on_submit():
+        user = User(username=form.username.data, email=form.email.data)
+        user.set_password(form.password.data)
+        db.session.add(user)
+        db.session.commit()
+        flash('Congratulations, you are now a registered user!')
+        return redirect(url_for('login'))
+    return render_template('register.html', form=form)
+
+@app.route('/logout')
+def logout():
+    logout_user()
+    return redirect(url_for('index'))
+
+class LoginForm(FlaskForm):
+    username = StringField('Username', validators=[DataRequired()])
+    password = PasswordField('Password', validators=[DataRequired()])
+    remember_me = BooleanField('Remember Me')
+    submit = SubmitField('Sign In')
+
+class RegistrationForm(FlaskForm):
+    username = StringField('Username', validators=[DataRequired()])
+    email = StringField('Email', validators=[DataRequired(), Email()])
+    password = PasswordField('Password', validators=[DataRequired()])
+    password2 = PasswordField(
+        'Repeat Password', validators=[DataRequired(), EqualTo('password')])
+    submit = SubmitField('Register')
+
+    def validate_username(self, username):
+        user = User.query.filter_by(username=username.data).first()
+        if user is not None:
+            raise ValidationError('Please use a different username.')
+
+    def validate_email(self, email):
+        user = User.query.filter_by(email=email.data).first()
+        if user is not None:
+            raise ValidationError('Please use a different email address.')
 
 class MyForm(FlaskForm):
 	body = StringField('Body', validators=[DataRequired()])
@@ -47,20 +118,35 @@ class TaskForm(FlaskForm):
 @app.route('/delete_task/<int:task_id>')
 def delete_task(task_id):
     task = Task.query.get(task_id)
-    db.session.delete(task)
-    db.session.commit()
+    if current_user == task.author:
+        db.session.delete(task)
+        db.session.commit()
     return redirect(request.referrer or url_for('index'))
 
 @app.route('/change_task_completion/<int:task_id>')
 def change_task_completion(task_id):
     task = Task.query.get(task_id)
-    if task.complete:
-    	task.complete = False
-    else:
-    	task.complete = True
-    task.last_updated = datetime.now()
-    db.session.commit()
+    if current_user == task.author:
+        if task.complete:
+        	task.complete = False
+        else:
+        	task.complete = True
+        task.last_updated = datetime.now()
+        db.session.commit()
     return redirect(request.referrer or url_for('index'))
+
+class User(UserMixin, db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    username = db.Column(db.String(64), index=True, unique=True)
+    email = db.Column(db.String(120), index=True, unique=True)
+    password_hash = db.Column(db.String(128))
+    tasks = db.relationship('Task', backref='author', lazy='dynamic')
+    def set_password(self, password):
+        self.password_hash = generate_password_hash(password)
+    def check_password(self, password):
+        return check_password_hash(self.password_hash, password)
+    def __repr__(self):
+        return '<User {}>'.format(self.username)
 
 class Task(db.Model):
     id = db.Column(db.Integer, primary_key=True)
@@ -69,10 +155,11 @@ class Task(db.Model):
     complete = db.Column(db.Boolean, default=False)
     last_updated = db.Column(db.DateTime, default=datetime.now)
     deadline = db.Column(db.String(140))
+    user_id = db.Column(db.Integer, db.ForeignKey('user.id'))
 
     def __repr__(self):
         return '<Task {}>'.format(self.body)
 
 @app.shell_context_processor
 def make_shell_context():
-    return {'db': db, 'Task': Task}
+    return {'db': db, 'Task': Task, 'User': User}
